# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the REST API for Aucto's marketplace, it is currently in v1.",
        "title": "AUCTO Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "*",
    "basePath": "/api/v1",
    "paths": {
...
        "/products": {
            "get": {
                "description": "Gets product information of products given query parameters provided in the Request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Gets Products with given query parameters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get products from a specific seller Id. Default is without any seller_id specified.",
                        "name": "seller_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort By a specific attribute of the product. Default is posted_date",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Get products by a specific product type, the types are 'Pre-Order' or 'Buy-Now'. Default is both will be selected",
                        "name": "product_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.GetProductResponseData"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.Message"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new product post with the supplied data, if the data is not valid it throws and error",
                "produces": [
                    "application/json"
                ],
                "summary": "Creates a new product post",
                "parameters": [
                    {
                        "description": "The Seller who posted the product's seller_id",
                        "name": "seller_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Title of the product",
                        "name": "title",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Short description of the product",
                        "name": "description",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Price as an int of the product",
                        "name": "price",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Condition of the product from a scale of 0 to 5",
                        "name": "condition",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Type of product sale: Buy-Now or Pre-Order",
                        "name": "product_type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Quantity of product to be put for sale",
                        "name": "product_quantity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/data.ProductCreateResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.Message"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Checks to see if a product with a given id exists and returns its product information if it does.",
                "produces": [
                    "application/json"
                ],
                "summary": "Gets a Product by its Product ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "product_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.GetProductResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.Message"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.Message"
                        }
                    }
                }
            }
        },
        "/products/{id}/images": {
            "post": {
                "description": "Adds images to an existing product with supplied product id. If product with product id does not exist returns a",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Adds images to products",
                "parameters": [
                    {
                        "type": "string",
                        "description": "product_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Array of image files to add to the product post",
                        "name": "images",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/data.Message"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.Message"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "$ref": "#/definitions/data.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.GetProductResponseData": {
            "type": "object",
            "required": [
                "condition",
                "desc",
                "images",
                "posted_date",
                "price",
                "product_id",
                "product_quantity",
                "product_type",
                "seller_info",
                "sold_quantity",
                "title"
            ],
            "properties": {
                "condition": {
                    "type": "integer"
                },
                "desc": {
                    "type": "string"
                },
                "discount": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.ProductImageData"
                    }
                },
                "order_by": {
                    "type": "string"
                },
                "posted_date": {
                    "type": "string",
                    "example": "2023-08-03 02:50:26.034552906 +0000 UTC m=+192.307467936"
                },
                "price": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "string"
                },
                "product_quantity": {
                    "type": "integer"
                },
                "product_type": {
                    "type": "string"
                },
                "releases_on": {
                    "type": "string"
                },
                "seller_info": {
                    "$ref": "#/definitions/data.GetSellerByIdResponseData"
                },
                "sold_quantity": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "data.GetSellerByIdResponseData": {
            "type": "object",
            "required": [
                "followers",
                "seller_id",
                "seller_name"
            ],
            "properties": {
                "followers": {
                    "type": "integer"
                },
                "seller_id": {
                    "type": "string"
                },
                "seller_name": {
                    "type": "string"
                }
            }
        },
        "data.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "data.ProductCreateResponseData": {
            "type": "object",
            "required": [
                "condition",
                "desc",
                "posted_date",
                "price",
                "product_id",
                "product_quantity",
                "product_type",
                "seller_id",
                "sold_quantity",
                "title"
            ],
            "properties": {
                "condition": {
                    "type": "integer"
                },
                "desc": {
                    "type": "string"
                },
                "discount": {
                    "type": "integer"
                },
                "order_by": {
                    "type": "string"
                },
                "posted_date": {
                    "type": "string",
                    "example": "2023-08-03 02:50:26.034552906 +0000 UTC m=+192.307467936"
                },
                "price": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "string"
                },
                "product_quantity": {
                    "type": "integer"
                },
                "product_type": {
                    "type": "string"
                },
                "releases_on": {
                    "type": "string"
                },
                "seller_id": {
                    "type": "string"
                },
                "sold_quantity": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "data.ProductImageData": {
            "type": "object",
            "required": [
                "image_no",
                "image_path"
            ],
            "properties": {
                "image_no": {
                    "type": "integer"
                },
                "image_path": {
                    "type": "string"
                }
            }
        },
...
    }
}